Fixed a bug in the way we handle if..then.else blocks
A bug has been identified and fixed in the way if..then..else
blocks were handled: the instruction right after the if block
had a context with `closebracket` and `ifblock` set, while
it was not part of the if block. If any other flow lead to
this instruction like it was jumping inside an if block in
its `then` or `else` branch it caused a flow analysis error
and a lot of trouble for Ghidra to follow what was going on
with the p-code (inconsistency in the execution flow reported
by Ghidra).

Therefore, we marked `closebracket` as `noflow` and modified
our definitions to set `ifblock` to 0 as soon as possible and
then exclude the instruction where we insert the closing bracket
from being considered as part of the if block.